Convert python integers into int32/int64 only during frame creation (#2127)
Before this PR, `dt.Frame([1, 2, 3])` would create an `int8` column, and `dt.Frame([100, 200, 300])` had stype `int16`. After this PR both examples will create an `int32` column instead.

The reason for this change is that it is too error-prone: the user may not notice that int8/int16 columns are created, and cause an integer overflow during simple mathematical operations. With int32 the likelihood of an overflow is much less.

+++ tests/munging/test_dt_rows.py
@@ -76,7 +76,7 @@ def test_dt0_properties(dt0):
     assert dt0.shape == (10, 3)  # must be a tuple, not a list!
     assert dt0.names == ("colA", "colB", "colC")
     assert dt0.ltypes == (ltype.bool, ltype.int, ltype.real)
-    assert dt0.stypes == (stype.bool8, stype.int16, stype.float64)
+    assert dt0.stypes == (stype.bool8, stype.int32, stype.float64)
     assert not any(frame_columns_virtual(dt0))
     frame_integrity_check(dt0)
